+++ rpc/backend/call_tx_test.go
@@ -24,6 +24,7 @@ func (suite *BackendTestSuite) TestResend() {
 	toAddr := tests.GenerateAddress()
 	chainID := (*hexutil.Big)(suite.backend.chainID)
 	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
+	height := int64(1)
 	callArgs := evmtypes.TransactionArgs{
 		From:                 nil,
 		To:                   &toAddr,
@@ -65,8 +66,8 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFeeDisabled(queryClient)
 				RegisterValidatorAccount(queryClient, validator)
@@ -87,10 +88,10 @@ func (suite *BackendTestSuite) TestResend() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				feeMarketClient := suite.backend.queryClient.FeeMarket.(*mocks.FeeMarketQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterFeeMarketParams(feeMarketClient, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterFeeMarketParams(feeMarketClient, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -108,9 +109,9 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFeeDisabled(queryClient)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -145,8 +146,8 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlockError(client, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeaderError(client, &height)
 			},
 			evmtypes.TransactionArgs{
 				Nonce: &txNonce,
@@ -162,9 +163,9 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -186,9 +187,9 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -208,12 +209,12 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterEstimateGas(queryClient, callArgs)
-				RegisterParams(queryClient, &header, 1)
-				RegisterParamsWithoutHeader(queryClient, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterParamsWithoutHeader(queryClient, height)
 				RegisterUnconfirmedTxsError(client, nil)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -237,12 +238,12 @@ func (suite *BackendTestSuite) TestResend() {
 				var header metadata.MD
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterEstimateGas(queryClient, callArgs)
-				RegisterParams(queryClient, &header, 1)
-				RegisterParamsWithoutHeader(queryClient, 1)
+				RegisterParams(queryClient, &header, height)
+				RegisterParamsWithoutHeader(queryClient, height)
 				RegisterUnconfirmedTxsEmpty(client, nil)
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -405,7 +406,8 @@ func (suite *BackendTestSuite) TestDoCall() {
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, bz)
+				height := int64(1)
+				RegisterHeader(client, &height, bz)
 				RegisterEthCallError(queryClient, &evmtypes.EthCallRequest{Args: argsBz, ChainId: suite.backend.chainID.Int64()})
 			},
 			rpctypes.BlockNumber(1),
@@ -418,7 +420,8 @@ func (suite *BackendTestSuite) TestDoCall() {
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, bz)
+				height := int64(1)
+				RegisterHeader(client, &height, bz)
 				RegisterEthCall(queryClient, &evmtypes.EthCallRequest{Args: argsBz, ChainId: suite.backend.chainID.Int64()})
 			},
 			rpctypes.BlockNumber(1),
@@ -447,7 +450,7 @@ func (suite *BackendTestSuite) TestDoCall() {
 func (suite *BackendTestSuite) TestGasPrice() {
 	defaultGasPrice := (*hexutil.Big)(big.NewInt(1))
 	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
-
+	height := int64(1)
 	testCases := []struct {
 		name         string
 		registerMock func()
@@ -461,10 +464,10 @@ func (suite *BackendTestSuite) TestGasPrice() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				feeMarketClient := suite.backend.queryClient.FeeMarket.(*mocks.FeeMarketQueryClient)
-				RegisterFeeMarketParams(feeMarketClient, 1)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterFeeMarketParams(feeMarketClient, height)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, sdkmath.NewInt(1))
 				RegisterValidatorAccount(queryClient, validator)
 			},
@@ -478,10 +481,10 @@ func (suite *BackendTestSuite) TestGasPrice() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				feeMarketClient := suite.backend.queryClient.FeeMarket.(*mocks.FeeMarketQueryClient)
-				RegisterFeeMarketParamsError(feeMarketClient, 1)
-				RegisterParams(queryClient, &header, 1)
-				RegisterBlock(client, 1, nil)
-				RegisterBlockResults(client, 1)
+				RegisterFeeMarketParamsError(feeMarketClient, height)
+				RegisterParams(queryClient, &header, height)
+				RegisterHeader(client, &height, nil)
+				RegisterBlockResults(client, height)
 				RegisterBaseFee(queryClient, sdkmath.NewInt(1))
 				RegisterValidatorAccount(queryClient, validator)
 			},