+++ rpc/backend/client_test.go
@@ -289,6 +289,29 @@ func RegisterHeaderByHashError(client *mocks.Client, hash common.Hash, tx []byte
 		Return(nil, errortypes.ErrInvalidRequest)
 }
 
+func RegisterHeaderByHashNotFound(client *mocks.Client, hash common.Hash, tx []byte) {
+	client.On("HeaderByHash", rpc.ContextWithHeight(1), bytes.HexBytes(hash.Bytes())).
+		Return(&tmrpctypes.ResultHeader{Header: nil}, nil)
+}
+
+// Header
+func RegisterHeader(client *mocks.Client, height *int64, tx []byte) (*tmrpctypes.ResultHeader, error) {
+	block := types.MakeBlock(*height, []types.Tx{tx}, nil, nil)
+	resHeader := &tmrpctypes.ResultHeader{Header: &block.Header}
+	client.On("Header", rpc.ContextWithHeight(*height), height).Return(resHeader, nil)
+	return resHeader, nil
+}
+
+func RegisterHeaderError(client *mocks.Client, height *int64) {
+	client.On("Header", rpc.ContextWithHeight(*height), height).Return(nil, errortypes.ErrInvalidRequest)
+}
+
+// Header not found
+func RegisterHeaderNotFound(client *mocks.Client, height int64) {
+	client.On("Header", rpc.ContextWithHeight(height), mock.AnythingOfType("*int64")).
+		Return(&tmrpctypes.ResultHeader{Header: nil}, nil)
+}
+
 func RegisterABCIQueryWithOptions(client *mocks.Client, height int64, path string, data bytes.HexBytes, opts tmrpcclient.ABCIQueryOptions) {
 	client.On("ABCIQueryWithOptions", context.Background(), path, data, opts).
 		Return(&tmrpctypes.ResultABCIQuery{