+++ tests/integration_tests/test_gas.py
@@ -119,4 +119,4 @@ def discard(request, tmp_path_factory):
 def test_discard_abci_responses(discard):
     with pytest.raises(ValueError) as exc:
         discard.w3.eth.gas_price
-    assert "block result not found for height" in str(exc)
+    assert "header result not found for height" in str(exc)